Move Docker OOM status tests to integration target
